Check categories in extmetadata against the filter as well
This almost caused something really awful to be tweeted, I'm glad
I caught this in time.

It appears that the list of categories you get when asking the API for
"categories" is not the same list that is shown in the article, for
some weird reason, and there's actually another list of categories
under extmetadata.