Check file usage agains the blacklist, and add more words to the blacklist
It feels awful having to write all these words together, but I need a very
good blacklist to make sure nothing racist will be tweeted by the bot.

Checking the file useage in various wikis gives yet another hint about
the nature of the file and if it's usable. Adding more heuristics like this,
even when they seem redundant at times, helps reduce the chance of the bot
posting something awful.

I keep thinking about this article in the context of picdescbot:
https://motherboard.vice.com/read/how-to-make-a-not-racist-bot

I'm using a dataset which I have not vetted myself (because I simply can't
vet every single picture on Commons), so adding a *lot* of filters is a must.