Filter out pictures based on tags returned from cvapi
This includes screenshots, pictures of text (or maps, the "giraffes
next to a map" joke is not funny anymore), and pictures of people in
military uniforms.

My main reasons for blacklisting pictures of text is not that the joke
got old, but rather the pictures themselves are boring, and also
there's a chance the text (especially if it's old) is saying something
really shitty.