Update the tumblr post template to workaround a tumblr bug
The tumblr app has an annoying bug in which it doesn't know what to
do with links to "pages" in a tumblr blog, and instead shows the
blogs front page.

Using a goo.gl redirect is a hacky way to work-around this bug,
because it makes the app think it's an external link, so it'll
open it in the browser, which will work.