Minor changes to adapt for the API moving to Azure
Project Oxford is dead. The old keys and endpoints for it no longer
work. But Azure has the same API, with configurable endpoints,
so picdescbot will live on.

This commit adds an option to configure the endpoint in the config
file, since the Azure computer vision API has multiple endpoints.