blacklist ultrasound images
The API seems to always call these "a black and white photo of a
person", even when it's something like a spleen or some other
internal organ.

I think this is due to some religious bias in Microsoft's training
set. Blacklisting ultrasound images is the best way to avoid this
damaging bias and any controversy it might cause.